requirements.txt
Pillow==9.4.0
requests==2.28.2
numpy==1.24.2
ultralytics==8.0.154
opencv-python==4.7.0.68
PyAutoGUI==0.9.54